requirements.txt
nextcord
ollama
python-dotenv
